Add invalid range rule to the list
