+++ scan-command/src/main/java/io/ballerina/scan/internal/CoreRule.java
@@ -20,7 +20,6 @@ package io.ballerina.scan.internal;
 
 import io.ballerina.scan.Rule;
 import io.ballerina.scan.RuleKind;
-import io.ballerina.scan.utils.Constants;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -34,8 +33,7 @@ enum CoreRule {
     AVOID_CHECKPANIC(RuleFactory.createRule(1, "Avoid checkpanic", RuleKind.CODE_SMELL)),
     UNUSED_FUNCTION_PARAMETER(RuleFactory.createRule(2,
             "Unused function parameter", RuleKind.CODE_SMELL)),
-    INVALID_RANGE_OPERATOR(RuleFactory.createRule(10,
-            Constants.RuleDescription.INVALID_RANGE_OPERATOR, RuleKind.CODE_SMELL));
+    INVALID_RANGE_OPERATOR(RuleFactory.createRule(10, "Invalid range operator", RuleKind.CODE_SMELL));
 
     private final Rule rule;
