+++ scan-command/src/test/java/io/ballerina/scan/internal/StaticCodeAnalyzerTest.java
@@ -27,7 +27,6 @@ import io.ballerina.scan.Issue;
 import io.ballerina.scan.Rule;
 import io.ballerina.scan.RuleKind;
 import io.ballerina.scan.Source;
-import io.ballerina.scan.utils.Constants;
 import io.ballerina.tools.text.LineRange;
 import org.testng.Assert;
 
@@ -45,35 +44,6 @@ public class StaticCodeAnalyzerTest extends BaseTest {
         Project project = SingleFileProject.load(coreRuleBalFiles.resolve(documentName));
         Module defaultModule = project.currentPackage().getDefaultModule();
         return defaultModule.document(defaultModule.documentIds().iterator().next());
-    }
-
-    @Test(description = "test checkpanic analyzer")
-    void testRangeOperatorAnalyzer() {
-        String documentName = "invalid_range_operator.bal";
-        Document document = loadDocument(documentName);
-        ScannerContextImpl scannerContext = new ScannerContextImpl(List.of(CoreRule.INVALID_RANGE_OPERATOR.rule()));
-        StaticCodeAnalyzer staticCodeAnalyzer = new StaticCodeAnalyzer(document, scannerContext);
-        staticCodeAnalyzer.analyze();
-        List<Issue> issues = scannerContext.getReporter().getIssues();
-        Assert.assertEquals(issues.size(), 9);
-        assertIssue(issues.get(0), documentName, 9, 21, 9, 27, "ballerina:10", 10,
-                Constants.RuleDescription.INVALID_RANGE_OPERATOR, RuleKind.CODE_SMELL);
-        assertIssue(issues.get(1), documentName, 13, 21, 13, 27, "ballerina:10", 10,
-                Constants.RuleDescription.INVALID_RANGE_OPERATOR, RuleKind.CODE_SMELL);
-        assertIssue(issues.get(2), documentName, 17, 21, 17, 28, "ballerina:10", 10,
-                Constants.RuleDescription.INVALID_RANGE_OPERATOR, RuleKind.CODE_SMELL);
-        assertIssue(issues.get(3), documentName, 41, 21, 41, 27, "ballerina:10", 10,
-                Constants.RuleDescription.INVALID_RANGE_OPERATOR, RuleKind.CODE_SMELL);
-        assertIssue(issues.get(4), documentName, 45, 21, 45, 27, "ballerina:10", 10,
-                Constants.RuleDescription.INVALID_RANGE_OPERATOR, RuleKind.CODE_SMELL);
-        assertIssue(issues.get(5), documentName, 49, 21, 49, 28, "ballerina:10", 10,
-                Constants.RuleDescription.INVALID_RANGE_OPERATOR, RuleKind.CODE_SMELL);
-        assertIssue(issues.get(6), documentName, 53, 21, 53, 26, "ballerina:10", 10,
-                Constants.RuleDescription.INVALID_RANGE_OPERATOR, RuleKind.CODE_SMELL);
-        assertIssue(issues.get(7), documentName, 57, 21, 57, 28, "ballerina:10", 10,
-                Constants.RuleDescription.INVALID_RANGE_OPERATOR, RuleKind.CODE_SMELL);
-        assertIssue(issues.get(8), documentName, 61, 21, 61, 28, "ballerina:10", 10,
-                Constants.RuleDescription.INVALID_RANGE_OPERATOR, RuleKind.CODE_SMELL);
     }
 
     void assertIssue(Issue issue, String documentName, int startLine, int startOffset, int endLine, int endOffset,
