+++ scan-command/src/test/resources/test-resources/core-rules/rule006_invalid_range_expression.bal
